Move SignalTerminationGuard to util
